Handle lazy finger touch.
When a finger lays over a soft-button, the other finger can be used to move the cursor. In this case, scrolling is prevented.

This patch only works when the lazy finger touches the soft-button *before* the other button. In the opposite case, however, a scrolling is still generated since it is not possible, with semi-multitouch devices, to discriminate the two fingers positions.